requirements.txt
pandas
plotly-express
statistics
